[backend] refactor: move organizations and users modules into organization-management
