+++ apps/portal-api/src/crons.ts
@@ -1,7 +1,7 @@
 import cron, { ScheduledTask } from 'node-cron';
 import { requestContext } from './context/request.context';
 import { DeploymentApp } from './modules/deployment/deployment.app';
-import { UsersOrganizationApp } from './modules/users/users.organization.app';
+import { UsersOrganizationApp } from './modules/organization-management/users/users.organization.app';
 import { EpicApp } from './modules/xtm-suite-roadmap/epic.app';
 import { SYSTEM_USER_CONTEXT } from './portal.const';
 import { logApp } from './utils/app-logger.util';
