+++ apps/portal-api/src/modules/deployment/group/service-group.app.ts
@@ -16,10 +16,10 @@ import { auth0Client } from '../../../thirdparty/auth0/client';
 import { logApp } from '../../../utils/app-logger.util';
 import { ErrorCode } from '../../../utils/error/error.code';
 import { formatName } from '../../../utils/format';
-import { organizationDomain } from '../../organizations/organizations.domain';
+import { organizationDomain } from '../../organization-management/organizations/organizations.domain';
+import { UsersDomain } from '../../organization-management/users/users.domain';
 import { PlatformConfiguration } from '../../registration/registration.domain';
 import { ServiceConfigurationDomain } from '../../registration/service-configuration/service-configuration.domain';
-import { UsersDomain } from '../../users/users.domain';
 import { DeploymentRequestDomain } from '../deployment.domain';
 import { ServiceGroupDomain } from './service-group.domain';
 import { ServiceGroupHelper } from './service-group.helper';
