+++ apps/portal-api/src/modules/organization-management/organizations/organizations.app.ts
@@ -1,15 +1,15 @@
 import { v4 as uuidv4 } from 'uuid';
-import { OrganizationInput } from '../../__generated__/resolvers-types';
-import { requestContext } from '../../context/request.context';
-import { OrganizationId } from '../../model/kanel/public/Organization';
-import { dispatch } from '../../pub';
-import { logApp } from '../../utils/app-logger.util';
-import { ErrorCode } from '../../utils/error/error.code';
-import { telemetryApp } from '../telemetry/telemetry.app';
+import { OrganizationInput } from '../../../__generated__/resolvers-types';
+import { requestContext } from '../../../context/request.context';
+import { OrganizationId } from '../../../model/kanel/public/Organization';
+import { dispatch } from '../../../pub';
+import { logApp } from '../../../utils/app-logger.util';
+import { ErrorCode } from '../../../utils/error/error.code';
+import { telemetryApp } from '../../telemetry/telemetry.app';
 import {
   buildCreateOrganizationEvent,
   buildUpdateOrganizationEvent,
-} from '../telemetry/telemetry.helper';
+} from '../../telemetry/telemetry.helper';
 import {
   deleteOrganizationBy,
   insertNewOrganization,
