+++ apps/portal-api/src/modules/organization-management/organizations/organizations.graphql
