+++ apps/portal-api/src/modules/organization-management/organizations/organizations.resolver.ts
@@ -1,8 +1,8 @@
-import { Resolvers } from '../../__generated__/resolvers-types';
-import { OrganizationId } from '../../model/kanel/public/Organization';
-import { UnknownErrorCode } from '../../utils/error/error.code';
-import { mapToGraphQLError } from '../../utils/error/error.mapping';
-import { StillReferencedError } from '../../utils/error/error.util';
+import { Resolvers } from '../../../__generated__/resolvers-types';
+import { OrganizationId } from '../../../model/kanel/public/Organization';
+import { UnknownErrorCode } from '../../../utils/error/error.code';
+import { mapToGraphQLError } from '../../../utils/error/error.mapping';
+import { StillReferencedError } from '../../../utils/error/error.util';
 import { organizationsApp } from './organizations.app';
 import {
   loadOrganizationBy,
