+++ apps/portal-api/src/modules/organization-management/users/users.auth.app.ts
@@ -1,8 +1,8 @@
-import { MutationLoginArgs } from '../../__generated__/resolvers-types';
-import { PORTAL_COOKIE_NAME } from '../../index';
-import { PortalContext } from '../../model/portal-context';
-import { UserLoadUserBy } from '../../model/user';
-import { validatePassword } from '../../security/utils/user';
+import { MutationLoginArgs } from '../../../__generated__/resolvers-types';
+import { PORTAL_COOKIE_NAME } from '../../../index';
+import { PortalContext } from '../../../model/portal-context';
+import { UserLoadUserBy } from '../../../model/user';
+import { validatePassword } from '../../../security/utils/user';
 import { loadUserBy, updateUserAtLogin } from './users.domain';
 
 const validPassword = (user: UserLoadUserBy, password: string): boolean => {
