+++ apps/portal-api/src/modules/organization-management/users/users.resolver.ts
@@ -4,19 +4,19 @@ import {
   User,
   UserPendingSubscription,
   UserSubscription,
-} from '../../__generated__/resolvers-types';
-import { OrganizationId } from '../../model/kanel/public/Organization';
-import { UserId } from '../../model/kanel/public/User';
-import { dispatch, listen } from '../../pub';
-import { hubspotReachOutSalesHook } from '../../thirdparty/hubspot/hubspot';
-import { logApp } from '../../utils/app-logger.util';
+} from '../../../__generated__/resolvers-types';
+import { OrganizationId } from '../../../model/kanel/public/Organization';
+import { UserId } from '../../../model/kanel/public/User';
+import { dispatch, listen } from '../../../pub';
+import { hubspotReachOutSalesHook } from '../../../thirdparty/hubspot/hubspot';
+import { logApp } from '../../../utils/app-logger.util';
 
-import { UserTransferRequestId } from '../../model/kanel/public/UserTransferRequest';
-import { PortalContext } from '../../model/portal-context';
-import { ErrorCode, UnknownErrorCode } from '../../utils/error/error.code';
-import { mapToGraphQLError } from '../../utils/error/error.mapping';
-import { ForbiddenAccess } from '../../utils/error/error.util';
-import { extractId } from '../../utils/utils';
+import { UserTransferRequestId } from '../../../model/kanel/public/UserTransferRequest';
+import { PortalContext } from '../../../model/portal-context';
+import { ErrorCode, UnknownErrorCode } from '../../../utils/error/error.code';
+import { mapToGraphQLError } from '../../../utils/error/error.mapping';
+import { ForbiddenAccess } from '../../../utils/error/error.util';
+import { extractId } from '../../../utils/utils';
 import { UserOrganizationPendingDomain } from './users-pending/user-organization-pending.domain';
 import { usersAdminApp } from './users.admin.app';
 import { UsersAuthApp } from './users.auth.app';
