+++ apps/portal-api/src/modules/registration/registration.domain.test.ts
@@ -30,7 +30,7 @@ import { PortalContext } from '../../model/portal-context';
 import { securityGuard } from '../../security/guard';
 import { ErrorCode } from '../../utils/error/error.code';
 import { DeploymentRequestDomain } from '../deployment/deployment.domain';
-import * as organizationDomain from '../organizations/organizations.domain';
+import * as organizationDomain from '../organization-management/organizations/organizations.domain';
 import { deleteServiceInstanceBy } from '../services/service-instance.domain';
 import * as subscriptionDomain from '../subcription/subscription.domain';
 import {
