+++ apps/portal-api/src/modules/services/service-instance.domain.ts
@@ -30,11 +30,11 @@ import { restrictSubscriptionToUserOrganization } from '../../security/restricti
 import { buildServiceLink, sendMail } from '../../server/mail-service';
 import { ServiceIdentifierToMailTemplate } from '../../server/mail-template/mail';
 import { formatRawObject } from '../../utils/queryRaw.util';
+import { loadUserBy } from '../organization-management/users/users.domain';
 import { PlatformConfiguration } from '../registration/registration.domain';
 import { loadSubscriptionWithOrganizationAndCapabilitiesBy } from '../subcription/subscription.helper';
 import { loadSubscriptionCapabilities } from '../user_service/service-capability/subscription-capability.domain';
 import { UserServiceDomain } from '../user_service/user_service.domain';
-import { loadUserBy } from '../users/users.domain';
 import { insertServiceCapability } from './instances/service-capabilities/service_capabilities.helper';
 
 export const ServiceInstanceDomain = {
