+++ apps/portal-api/src/modules/subcription/subscription.domain.ts
@@ -11,10 +11,10 @@ import Subscription, {
 } from '../../model/kanel/public/Subscription';
 import { UserMutator } from '../../model/kanel/public/User';
 import { restrictSubscriptionToUserOrganization } from '../../security/restriction/user-service';
-import { loadOrganizationBy } from '../organizations/organizations.domain';
+import { loadOrganizationBy } from '../organization-management/organizations/organizations.domain';
+import { loadUserBy } from '../organization-management/users/users.domain';
 import { loadServiceInstanceBy } from '../services/service-instance.domain';
 import { UserServiceDomain } from '../user_service/user_service.domain';
-import { loadUserBy } from '../users/users.domain';
 import { loadSubscriptionWithOrganizationAndCapabilitiesBy } from './subscription.helper';
 
 export const SubscriptionDomain = {
