+++ apps/portal-api/src/server/graphql-schema.ts
@@ -7,7 +7,8 @@ import deploymentResolver from '../modules/deployment/deployment.resolver';
 import serviceGroupResolver from '../modules/deployment/group/service-group.resolver';
 import ingestManifestResolver from '../modules/ingest-manifest/ingest-manifest.resolver';
 import logResolver from '../modules/log/log.resolver';
-import organizationsResolver from '../modules/organizations/organizations.resolver';
+import organizationsResolver from '../modules/organization-management/organizations/organizations.resolver';
+import usersResolver from '../modules/organization-management/users/users.resolver';
 import registrationResolver from '../modules/registration/registration.resolver';
 import ServiceDefinitionResolver from '../modules/services/definition/service-definition.resolver';
 import vaultResolver from '../modules/services/document/document.resolver';
@@ -21,7 +22,6 @@ import telemetryResolver from '../modules/telemetry/telemetry.resolver';
 import useCaseResolver from '../modules/use-case/use-case.resolver';
 import serviceCapabilityResolver from '../modules/user_service/service-capability/service-capability.resolver';
 import userServiceResolver from '../modules/user_service/user_service.resolver';
-import usersResolver from '../modules/users/users.resolver';
 import xtmSuiteRoadmapResolver from '../modules/xtm-suite-roadmap/epic.resolver';
 import nodesResolver from '../nodes/nodes.resolver';
 import { authDirectiveTransformer } from '../security/directive-graphql/directive-auth';
